fix(review): address PR review comments
